requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
accelerate==0.27.2
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
einops==0.8.2
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging 
peft==0.8.2
pillow==12.1.1
psutil==7.2.2
Pygments==2.19.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.33.0
rich==14.3.3
safetensors==0.7.0
shellingham==1.5.4
sympy==1.14.0
timm==1.0.26
tokenizers==0.15.2
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
transformers==4.38.2
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
